+++ lib/cubit/transfer/form/transfer_form_cubit.dart
@@ -42,7 +42,7 @@ class TransferFormCubit extends Cubit<TransferFormState> {
     ));
   }
 
-  void updateMoney(double money) {
+  void updateMoney(int money) {
     emit(
       state.copyWith(
         userInfo: state.userInfo.copyWith(money: money),
